refactor: apply ktlint
