+++ src/main/kotlin/com/freewill/security/oauth/key/OidcPublicKeyResponse.kt
@@ -1,4 +1,4 @@
-package com.freewill.security.oauth.dto.key
+package com.freewill.security.oauth.key
 
 import java.util.function.Supplier
 import javax.naming.AuthenticationException
